Application layout: Refactor core tool functions to separate module
