+++ pyproject.toml
@@ -74,7 +74,7 @@ dependencies = [
   "cachetools<7",
   "click<9",
   "cratedb-about==0.0.5",
-  "fastmcp<2.7",
+  "fastmcp>=2.7,<2.10",
   "hishel<0.2",
   "pueblo==0.0.11",
   "sqlparse<0.6",
